feat: validate delete request with zod

+++ src/employees/infrastructure/controllers/delete-employee.controller.ts
@@ -1,5 +1,6 @@
 import { Controller } from "@/protocols/controller";
 import { HttpResponses } from "@/protocols/http-responses";
+import { deleteEmployeeValidator } from "../presentation/request/delete-employe.validation";
 import { DeleteEmployeeRepository } from "../repositories/delete-employee.repository";
 
 type ByIdParameter = {
@@ -11,7 +12,14 @@ export class DeleteEmployeeController implements Controller<ByIdParameter> {
 
   async execute(request: ByIdParameter) {
     try {
+      const validatedRequest = deleteEmployeeValidator(request);
+
+      if (!validatedRequest.success) {
+        return HttpResponses.BadRequest(validatedRequest.error.errors);
+      }
+
       const response = await this.repository.deleteById(request.id);
+
       if (!response) {
         return HttpResponses.NotFound();
       }
